src/test/java/com/mycompany/Gym/MaintenanceTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.mycompany.Gym;

import junit.framework.Assert;
import static org.testng.Assert.*;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author student
 */
public class MaintenanceTest {
    
    public MaintenanceTest() {
    }

    // TODO add test methods here.
    // The methods must be annotated with annotation @Test. For example:
    //
     @Test
     public void create() 
     {
         Maintenance obj =  new Maintenance.Builder("02/03/2014").cost(1423.4).build();
         Assert.assertEquals(1423.4, obj.getCosts());
     }
     
     @Test
     public void update() 
     {
         Maintenance obj =  new Maintenance.Builder("02/03/2014").cost(2423.4).build();
         Assert.assertEquals(2423.4, obj.getCosts());
     }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}
